Added possible move, roll/dash and attack states
The enemies are now Kinematic because when they weren't their AI
pathfinding stuffs got all messed up if the player walked into them.
Everything in the Scripts folder is new.

Err I'm at work now so when I get home I'll try to explain how it works
its pretty simple if you want to give it a look. The main theme is..
PlayerController gives inputs to the current state, the current state
takes that input and if it is valid input for that state it does
something, usually getting CharacterDetails class to do it, sometimes
it moves to a new state making the current state that one.
If the input is not valid it just doesn't do anyting.

The states I can think of right now are, Default, FlyBack, Recover,
Attack, Roll/Dash, DashAttack, Sprint, SprintAttack, Heal, Shoot,
Staggered.